refactor(app): extract CSV validation and status badges from RFQTab
Moves the condition CSV row validator (validateConditionCsvRow + types) into
~/lib/admin/conditionCsv.ts with a unit test suite, and lifts the inner
ConditionStatusBadges subcomponent (including its WRAPPED_MARKETS_ABI and
on-chain read) out to its own file.

RFQTab shrinks from 1396 -> 1232 lines. Pure refactor — no UI or behavior
changes.
